Silence day-20 warnings.
We're not using the code that computes the number of distinct ways a regex
can match, but it's kind of interesting to keep it around.

This overcounts the number of distinct matches, unfortunately. For example, a
regex like `(|A)(|A)(|A)` matches four paths: the empty path, `A`, `AA`, and
`AAA`. But matches those paths in eight different ways: each alternation could
either match or not match, independently of the others.